Refactord main.rs into lib.rs as Preperation for android support
