+++ examples/demo/Cargo.toml
@@ -6,6 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-bevy = "0.11.3"
+bevy = "0.12"
 bevy_openxr = "0.0.1"
 bevy_rapier3d = { git = "https://github.com/alexichepura/bevy_rapier", version = "0.22.0" }
